requirements.txt

# The base set of requirements for REST framework is actually
# just Django, but for the purposes of development and testing
# there are a number of packages that are useful to install.

# Laying these out as seperate requirements files, allows us to
# only included the relevent sets when running tox, and ensures
# we are only ever declaring our dependencies in one place.

-r requirements/requirements-mandatory.txt
-r requirements/requirements-optionals.txt
-r requirements/requirements-testing.txt
-r requirements/requirements-documentation.txt
-r requirements/requirements-codestyle.txt
-r requirements/requirements-packaging.txt

Django~=3.2
djangorestframework~=3.13.1